requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pydantic>=2.5.0
asyncio-throttle>=1.0.2
aiohttp>=3.9.0
python-dotenv>=1.0.0 